test: split out sqlite session test and fix null __proto__
